Update the core localisation sync scripts, workflows and instructions
