+++ libretro-common/samples/core_options/example_translation/.github/workflows/crowdin_initial_setup.yml
@@ -1,0 +1,32 @@
+# Uploads source texts and any present translations to Crowdin.
+
+name: Crowdin Translations Initial Setup
+
+# On manual run only.
+on:
+  workflow_dispatch
+
+jobs:
+  upload_source_file:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Setup Java JDK
+        uses: actions/setup-java@v4
+        with:
+          java-version: 18
+          distribution: zulu
+
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Initial synchroniation
+        shell: bash
+        env:
+          CROWDIN_API_KEY: ${{ secrets.CROWDIN_API_KEY }}
+        run: |
+          python3 intl/initial_sync.py $CROWDIN_API_KEY "<CORE_NAME>" "<PATH/TO>/libretro_core_options.h"
