+++ libretro-common/samples/core_options/example_translation/intl/.gitignore
@@ -1,0 +1,4 @@
+__pycache__
+crowdin-cli.jar
+*.h
+*.json
