--- libretro-common/samples/core_options/example_translation/translation scripts/intl/.gitignore
@@ -1,1 +1,0 @@
-__pycache__
